some minor changes in Customer model

+++ src/main/java/com/shop_center/hyper_market_ali/customer/CustomerRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
 public interface CustomerRepository
-        extends JpaRepository<Customer, Integer> {
+        extends JpaRepository<Customer, Long> {
 
     @Query("SELECT c FROM Customer c WHERE c.email = ?1")
     Optional<Customer> findCustomerByEmail(String email);
